+++ R/svyjskm.R
@@ -6,6 +6,7 @@
 #' @param ylabs y-axis label.
 #' @param xlims numeric: list of min and max for x-axis. Default: NULL
 #' @param ylims numeric: list of min and max for y-axis. Default: c(0, 1)
+#' @param surv.scale 	scale transformation of survival curves. Allowed values are "default" or "percent".
 #' @param ystratalabs character list. A list of names for each strata. Default: NULL
 #' @param ystrataname The legend name. Default: 'Strata'
 #' @param timeby numeric: control the granularity along the time-axis; defaults to 7 time-points. 
@@ -44,11 +45,12 @@
 
 svyjskm <- function(sfit,
                     xlabs = "Time-to-event",
-                    ylabs = "Survival (%)",
+                    ylabs = "Survival probability",
                     xlims = NULL,
                     ylims = c(0,1),
                     ystratalabs = NULL,
                     ystrataname = "Strata",
+                    surv.scale = c("default", "percent"),
                     timeby = NULL,
                     main = "",
                     pval = FALSE,
@@ -161,6 +163,12 @@ svyjskm <- function(sfit,
     linetype=c("solid", "solid", "solid", "solid", "solid", "solid", "solid", "solid", "solid", "solid", "solid")
   }
   
+  # Scale transformation
+  #::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
+  surv.scale <- match.arg(surv.scale)
+  scale_labels <-  ggplot2::waiver()
+  if (surv.scale == "percent") scale_labels <- scales::percent
+  
   
   p <- ggplot2::ggplot( df, aes(x=time, y=surv, colour=strata, linetype=strata)) + ggtitle(main)
   
@@ -178,7 +186,7 @@ svyjskm <- function(sfit,
           axis.line.x = element_line(size = 0.5, linetype = "solid", colour = "black"),
           axis.line.y = element_line(size = 0.5, linetype = "solid", colour = "black")) +
     scale_x_continuous(xlabs, breaks = times, limits = xlims) +
-    scale_y_continuous(ylabs, limits = ylims)
+    scale_y_continuous(ylabs, limits = ylims, labels = scale_labels)
   
   
   #Add 95% CI to plot