+++ tests/testthat/test-km.R
@@ -6,10 +6,10 @@ data(colon);data(pbc)
 test_that("Run jskm", {
   fit <- survfit(Surv(time,status)~rx, data=colon)
   expect_is(jskm(fit, timeby=500, table = T, pval = T), "gtable")
-  expect_is(jskm(fit, table = T, pval = T, label.nrisk = "No. at risk", timeby = 365, xlabs = "Time(Day)", ylabs = "Survival(%)", marks = F, xlims = c(0, 3500), ylims = c(0.25, 1),
+  expect_is(jskm(fit, table = T, pval = T, label.nrisk = "No. at risk", timeby = 365, xlabs = "Time(Day)", ylabs = "Survival", marks = F, xlims = c(0, 3500), ylims = c(0.25, 1),
        ystratalabs = c("Obs", "Lev", "Lev + 5FU"), ystrataname = "rx"), "gtable")
   expect_is(jskm(fit, timeby=500), "gg")
-  expect_is(jskm(fit, timeby=500, main = "kaplan", xlabs = "time", ylabs = "suvrival"), "gg")
+  expect_is(jskm(fit, timeby=500, main = "kaplan", xlabs = "time", ylabs = "Suvrival (%)", surv.scale = "percent"), "gg")
   expect_is(jskm(fit), "gg")
   expect_is(jskm(fit, timeby=500, ci = T), "gg")
   expect_is(jskm(fit, timeby=500, legend = F), "gg")
@@ -33,7 +33,7 @@ dpbc<-survey::svydesign(id=~1, prob=~randprob, strata=~edema, data=subset(pbc,ra
 test_that("Run svyjskm", {
   s1 <- survey::svykm(Surv(time,status>0)~sex, design=dpbc, se = T)
   expect_is(svyjskm(s1, ci = T), "gg")
-  expect_is(svyjskm(s1, ci = F), "gg")
+  expect_is(svyjskm(s1, ci = F, ylabs = "Suvrival (%)", surv.scale = "percent"), "gg")
   expect_is(svyjskm(s1, ci = T, cumhaz = T), "gg")
   expect_is(svyjskm(s1, ci = F, cumhaz = T), "gg")
   s2 <- survey::svykm(Surv(time,status>0)~sex, design=dpbc, se = F)